Adicionando testes reserva, acessoPermitido, almoxarifado, servico

+++ src/main/java/dados/Servico.java
@@ -9,7 +9,7 @@ public class Servico {
     @GeneratedValue
     private Integer id;
     private String descricao;
-    private Float valor;
+    private Double valor;
     private String codigo;
     @ManyToOne(targetEntity = Fornecedor.class, fetch = FetchType.LAZY)
     private Funcionario requerente;
@@ -19,7 +19,7 @@ public class Servico {
     public Servico() {
     }
 
-    public Servico(Integer id, String descricao, Float valor, String codigo, Funcionario requerente,
+    public Servico(Integer id, String descricao, Double valor, String codigo, Funcionario requerente,
             Fornecedor fornecedor) {
         this.id = id;
         this.descricao = descricao;
@@ -41,11 +41,11 @@ public class Servico {
         this.descricao = descricao;
     }
 
-    public Float getValor() {
+    public Double getValor() {
         return this.valor;
     }
 
-    public void setValor(Float valor) {
+    public void setValor(Double valor) {
         this.valor = valor;
     }
